S1-dlt/pyproject.toml
[project]
name = "s1-dlt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dlt[workspace]>=1.22.2",
    "ibis-framework[duckdb]>=12.0.0",
    "marimo>=0.20.2",
    "plotly>=6.6.0",
]
